requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3
#datascience
jupyter==1.0.0
pandas==1.5.0
numpy==1.23.3
scikit-learn==1.1.2
matplotlib==3.6.0
seaborn==0.12.0
dask==2022.9.2
Pillow==9.3.0
#MLOps
gradio==3.4.1
streamlit==1.13.0
protobuf==3.16.0
onnx==1.12.0
torchvision==0.14.0
